Update QA, created new file for data fetching
